requirements.txt
w1thermsensor
RPi.GPIO
awscrt
awsiot